move repo syncing code to the sync crate
