+++ crates/gitbutler-sync/src/cloud.rs
@@ -1,9 +1,12 @@
+use std::sync::atomic::AtomicUsize;
+use std::sync::Arc;
 use std::time;
 
-use anyhow::{Context, Result};
+use anyhow::{anyhow, Context, Result};
 use gitbutler_branchstate::VirtualBranchesAccess;
+use gitbutler_core::error::Code;
+use gitbutler_core::git::Url;
 use gitbutler_core::id::Id;
-use gitbutler_core::project_repository::RepoActions;
 use gitbutler_core::{
     git::{self},
     project_repository,
@@ -49,7 +52,7 @@ pub async fn sync_with_gitbutler(
         .map(|sha| format!("+{}:refs/gitbutler/oplog/oplog", sha));
 
     if let Some(oplog_refspec) = oplog_refspec {
-        let x = project_repository.push_to_gitbutler_server(Some(user), &[&oplog_refspec]);
+        let x = push_to_gitbutler_server(project_repository, Some(user), &[&oplog_refspec]);
         println!("\n\n\nHERE: {:?}", x?);
     }
 
@@ -82,7 +85,7 @@ async fn push_target(
     for (idx, id) in ids.iter().enumerate().rev() {
         let refspec = format!("+{}:refs/push-tmp/{}", id, project_id);
 
-        project_repository.push_to_gitbutler_server(Some(user), &[&refspec])?;
+        push_to_gitbutler_server(project_repository, Some(user), &[&refspec])?;
         update_project(projects, project_id, *id).await?;
 
         tracing::info!(
@@ -93,7 +96,8 @@ async fn push_target(
         );
     }
 
-    project_repository.push_to_gitbutler_server(
+    push_to_gitbutler_server(
+        project_repository,
         Some(user),
         &[&format!("+{}:refs/{}", default_target.sha, project_id)],
     )?;
@@ -167,7 +171,7 @@ fn push_all_refs(
 
     let all_refs: Vec<_> = all_refs.iter().map(String::as_str).collect();
 
-    let anything_pushed = project_repository.push_to_gitbutler_server(Some(user), &all_refs)?;
+    let anything_pushed = push_to_gitbutler_server(project_repository, Some(user), &all_refs)?;
     if anything_pushed {
         tracing::info!(
             %project_id,
@@ -193,4 +197,80 @@ async fn update_project(
         .await
         .context("failed to update last push")?;
     Ok(())
+}
+
+fn push_to_gitbutler_server(
+    project_repo: &project_repository::ProjectRepo,
+    user: Option<&users::User>,
+    ref_specs: &[&str],
+) -> Result<bool> {
+    let project = project_repo.project();
+    let url = project
+        .api
+        .as_ref()
+        .context("api not set")?
+        .code_git_url
+        .as_ref()
+        .context("code_git_url not set")?
+        .as_str()
+        .parse::<Url>()?;
+
+    tracing::debug!(
+        project_id = %project.id,
+        %url,
+        "pushing code to gb repo",
+    );
+
+    let user = user
+        .context("need user to push to gitbutler")
+        .context(Code::ProjectGitAuth)?;
+    let access_token = user.access_token()?;
+
+    let mut callbacks = git2::RemoteCallbacks::new();
+    if project.omit_certificate_check.unwrap_or(false) {
+        callbacks.certificate_check(|_, _| Ok(git2::CertificateCheckStatus::CertificateOk));
+    }
+    let bytes_pushed = Arc::new(AtomicUsize::new(0));
+    let total_objects = Arc::new(AtomicUsize::new(0));
+    {
+        let byte_counter = Arc::<AtomicUsize>::clone(&bytes_pushed);
+        let total_counter = Arc::<AtomicUsize>::clone(&total_objects);
+        callbacks.push_transfer_progress(move |_current, total, bytes| {
+            byte_counter.store(bytes, std::sync::atomic::Ordering::Relaxed);
+            total_counter.store(total, std::sync::atomic::Ordering::Relaxed);
+        });
+    }
+
+    let mut push_options = git2::PushOptions::new();
+    push_options.remote_callbacks(callbacks);
+    let auth_header = format!("Authorization: {}", access_token.0);
+    let headers = &[auth_header.as_str()];
+    push_options.custom_headers(headers);
+
+    let mut remote = project_repo.repo().remote_anonymous(&url.to_string())?;
+
+    remote
+        .push(ref_specs, Some(&mut push_options))
+        .map_err(|err| match err.class() {
+            git2::ErrorClass::Net => anyhow!("network failed"),
+            _ => match err.code() {
+                git2::ErrorCode::Auth => anyhow!("authentication failed")
+                    .context(Code::ProjectGitAuth)
+                    .context(err),
+                _ => anyhow!("push failed"),
+            },
+        })?;
+
+    let bytes_pushed = bytes_pushed.load(std::sync::atomic::Ordering::Relaxed);
+    let total_objects_pushed = total_objects.load(std::sync::atomic::Ordering::Relaxed);
+
+    tracing::debug!(
+        project_id = %project.id,
+        ref_spec = ref_specs.join(" "),
+        bytes = bytes_pushed,
+        objects = total_objects_pushed,
+        "pushed to gb repo tmp ref",
+    );
+
+    Ok(total_objects_pushed > 0)
 }
